+++ universal_ads_sdk/endpoints/__init__.py
@@ -7,6 +7,10 @@ from .media import MediaEndpoint
 from .report import ReportEndpoint
 from .segment import SegmentEndpoint
 from .me import MeEndpoint
+from .campaign import CampaignEndpoint
+from .adset import AdsetEndpoint
+from .ad import AdEndpoint
+from .pixel import PixelEndpoint
 
 __all__ = [
     "CreativeEndpoint",
@@ -14,4 +18,8 @@ __all__ = [
     "ReportEndpoint",
     "SegmentEndpoint",
     "MeEndpoint",
+    "CampaignEndpoint",
+    "AdsetEndpoint",
+    "AdEndpoint",
+    "PixelEndpoint",
 ]